feat: port number in check
The existing implementation only checks based on the hostname. If a uri links to
a different portnumber on the same host, this was not considered a redirect. In
situations like local development were multiple services might be running on the
same host, this results in links not working.

This commit includes the portnumber into the validation. In case no explicit
port is mentioned, both `port` and `current-port` will return nil. If an
explicit port is used, the comparison will be based on port numbers.

I have not verified the browser support of this feature, so I do not know if
this breaks any existing setups on non-mainstream browsers.