fix: port number
The fix proposed in https://github.com/venantius/accountant/pull/35 was too
rigorous. Relative URLs lack a port number, and thus a link on a domain with a
port triggers a reload. This is not efficient.

This fix resolves this issue by only executing the check in case a port number
is specified in the link. Other checks on the domain to ensure proper reloeads
are already cared for.

Ports are compared as strings, to avoid any type issues.